ai_engine/requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
vertexai
pandas
db-dtypes
asyncio
